define a models.py for the model to live in
